requirements.txt
python-dotenv==0.20.0
requests==2.28.1
sendgrid==6.9.7
boto3==1.24.58
